Initial version of MC6809E-to-NEXYS2 FPGA board adapter. PCB is being fabbed, so design is as yet untested.
One flaw I've already noticed: The FPGA board connector is on the wrong side of the board! This should be an easy revision for the next release. It may also be possible to very carefully modify rework the board to make this revision work -- power supply and ground traces on the A row will need to be cut and rewired. Not pretty, but probably doable.